refactor: update team metrics polling interval and clean up chart imports
- Adjusted `TEAM_METRICS_INITIAL_RANGE_MS` from 1 hour to 5 minutes for improved responsiveness.
- Refactored chart component imports to streamline file structure by renaming and removing legacy files (`concurrent-chart.client.tsx` and `start-rate-chart.client.tsx`).
