+++ src/configs/intervals.ts
@@ -2,7 +2,7 @@ export const SANDBOXES_METRICS_POLLING_MS = 3_000
 
 export const SANDBOXES_DETAILS_METRICS_POLLING_MS = 3_000
 
-export const TEAM_METRICS_INITIAL_RANGE_MS = 60 * 60 * 1000 // 1 hour
+export const TEAM_METRICS_INITIAL_RANGE_MS = 60 * 5 * 1_000 // 5 minutes
 
 export const TEAM_METRICS_POLLING_INTERVAL_MS = 10_000
